improvements in use-case file structure
