+++ src/use-cases/addresses/fetch-customer-address.ts
@@ -1,6 +1,6 @@
 import { AddressRepository } from '@/repositories/addresses-repository'
 import { Address } from '@prisma/client'
-import { ResourceNotFoundError } from './errors/resource-not-found-error'
+import { ResourceNotFoundError } from '../errors/resource-not-found-error'
 
 interface FetchCustomerAddressUseCaseRequest {
   customerId: number
