+++ src/use-cases/customers/authenticate.ts
@@ -1,5 +1,5 @@
 import { CustomerRepository } from '@/repositories/customers-repository'
-import { InvalidCredentialsError } from './errors/invalid-credentials-error'
+import { InvalidCredentialsError } from '../errors/invalid-credentials-error'
 import { compare } from 'bcryptjs'
 import { Customer } from '@prisma/client'
