+++ src/use-cases/factories/make-add-product-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaProductsRepository } from '@/repositories/prisma/prisma-products-repository'
-import { AddProductUseCase } from '../add-product'
+import { AddProductUseCase } from '@/use-cases/orders/add-product'
 
 export function makeAddProductUseCase() {
   const prismaProductRepository = new PrismaProductsRepository()
