+++ src/use-cases/factories/make-authenticate-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaCustomersRepository } from '@/repositories/prisma/prisma-customers-repository'
-import { AuthenticateUseCase } from '../authenticate'
+import { AuthenticateUseCase } from '@/use-cases/customers/authenticate'
 
 export function makeAuthenticateUseCase() {
   const prismaCustomerRepository = new PrismaCustomersRepository()
