+++ src/use-cases/factories/make-register-customer-address-use-case.ts
@@ -1,5 +1,5 @@
 import { PrismaAddressesRepository } from '@/repositories/prisma/prisma-addresses-repository'
-import { RegisterCustomerAddressUseCase } from '../register-customer-address'
+import { RegisterCustomerAddressUseCase } from '@/use-cases/addresses/register-customer-address'
 
 export function makeRegisterCustomerAddressUseCase() {
   const prismaCustomerAddressRepository = new PrismaAddressesRepository()
