+++ src/use-cases/orders/fetch-products.ts
@@ -1,5 +1,5 @@
 import { Product } from '@prisma/client'
-import { ResourceNotFoundError } from './errors/resource-not-found-error'
+import { ResourceNotFoundError } from '../errors/resource-not-found-error'
 import { ProductRepository } from '@/repositories/products-repository'
 
 interface FetchProductsUseCaseRequest {
